% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifttt.R
\name{convert_ifttt_time}
\alias{convert_ifttt_time}
\title{Convert itfff time to ymd-hm}
\usage{
convert_ifttt_time(chr)
}
\arguments{
\item{chr}{time in ifttt format}
}
\description{
Convert itfff time to ymd-hm
}
