% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\docType{data}
\name{jpn}
\alias{jpn}
\title{Use Japanese font}
\format{
An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 1.
}
\usage{
jpn
}
\description{
Use Japanese font
}
\keyword{datasets}
