% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_numbers.R
\name{lead_zero}
\alias{lead_zero}
\title{Add leading zero}
\usage{
lead_zero(x, digits = 2)
}
\arguments{
\item{x}{a vector of integer}

\item{digits}{number digits after zero padding}
}
\description{
Add leading zero
}
