% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cr.R
\name{read_cr}
\alias{read_cr}
\title{read function for Campbell logger}
\usage{
read_cr(
  file,
  n_max = Inf,
  wide = TRUE,
  remove_quality_check = TRUE,
  remove_time_of_occurence = TRUE
)
}
\description{
read function for Campbell logger
}
