% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_numbers.R
\name{tail_zero}
\alias{tail_zero}
\title{Add tailing zero}
\usage{
tail_zero(x, digits = 2)
}
\arguments{
\item{x}{a vector of numeric}

\item{digits}{number of decimal digits}
}
\description{
Add tailing zero
}
