% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_files.R
\name{write_missing_files}
\alias{write_missing_files}
\title{Write only files that are not available}
\usage{
write_missing_files(write_fun, x, file, ..., msg = c("skip:", "written:"))
}
\arguments{
\item{write_fun}{file-write function}

\item{x}{object to be written}

\item{file}{filename}
}
\description{
Write only files that are not available
}
