% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\docType{data}
\name{transparent}
\alias{transparent}
\title{Transparent backgrounds}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 4.
}
\usage{
transparent
}
\description{
Transparent backgrounds
}
\keyword{datasets}
